+++ Mindscape.Raygun4Net.WebApi.Tests/Model/FakeRaygunWebApiClient.cs
@@ -1,4 +1,5 @@
-﻿using Mindscape.Raygun4Net.Messages;
+﻿using Mindscape.Raygun4Net;
+using Mindscape.Raygun4Net.Messages;
 using System;
 using System.Collections.Generic;
 
@@ -6,6 +7,20 @@ namespace Mindscape.Raygun4Net.WebApi.Tests.Model
 {
   public class FakeRaygunWebApiClient : RaygunWebApiClient
   {
+    public FakeRaygunWebApiClient()
+    {
+    }
+
+    public FakeRaygunWebApiClient(string apiKey)
+      : base(apiKey)
+    {
+    }
+
+    public RaygunRequestMessageOptions ExposeRequestMessageOptions
+    {
+      get { return _requestMessageOptions; }
+    }
+
     public bool ExposeCanSend(RaygunMessage message)
     {
       return CanSend(message);